pyproject.toml
[tool.poetry]
name = "score-simulator-py"
version = "0.1.0"
description = "score simulator py"
authors = ["tanzhijian <tanzhijianorg@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.9.0"}
httpx = "^0.25.2"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mypy = "^1.7.1"
ruff = "^0.1.7"
respx = "^0.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88

[tool.mypy]
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests",]
